+++ docs/CHANGES.rst
@@ -2,11 +2,12 @@ Changes
 =======
 
 
-0.3b / (unreleased)
+0.3b / (2014-03-19)
 -------------------
 
-- Merged a fix from ixmatus (Parnell Springmeyer) so exception messages that aren't passed as args won't
-  cause trouble. - thanks a lot!
+- View groups don't catch exceptions anylonger. This was a cause of a lot of
+  odd error messages on Python 2.
+- Ordering with bad keys don't cause exceptions - they're logged as warnings instead.
 - New argument priority for view actions. [ixmatus]